requirements.txt
asgiref==3.3.1
boto3==1.17.50
botocore==1.20.50
Django==3.1.7
django-environ==0.4.5
django-storages==1.11.1
gunicorn==20.1.0
jmespath==0.10.0
Pillow==8.1.1
psycopg2==2.8.6
python-dateutil==2.8.1
python-http-client==3.3.2
pytz==2021.1
s3transfer==0.3.6
sendgrid==6.7.0
six==1.15.0
sqlparse==0.4.1
starkbank-ecdsa==1.1.0
urllib3==1.26.4
whitenoise==5.2.0
